+++ lm_eval/__main__.py
@@ -9,9 +9,8 @@ from typing import Union
 
 from lm_eval import evaluator, utils
 from lm_eval.evaluator import request_caching_arg_to_dict
-from lm_eval.logging_utils import WandbLogger
+from lm_eval.logging import EvaluationTracker, WandbLogger
 from lm_eval.tasks import TaskManager
-from lm_eval.tracking import EvaluationTracker
 from lm_eval.utils import handle_non_serializable, make_table, simple_parse_args_string
