+++ lm_eval/logging/__init__.py
@@ -1,1 +1,2 @@
 from .evaluation_tracker import EvaluationTracker
+from .wandb_logger import WandbLogger
